+++ app/config/events.php
@@ -95,6 +95,65 @@ return [
         '$model' => Response::MODEL_DATABASE,
         '$resource' => true,
         '$description' => 'This event triggers on any database event.',
+        'tables' => [
+            '$model' => Response::MODEL_TABLE,
+            '$resource' => true,
+            '$description' => 'This event triggers on any table event.',
+            'rows' => [
+                '$model' => Response::MODEL_ROW,
+                '$resource' => true,
+                '$description' => 'This event triggers on any rows event.',
+                'create' => [
+                    '$description' => 'This event triggers when a row is created.',
+                ],
+                'update' => [
+                    '$description' => 'This event triggers when a row is updated.'
+                ],
+                'upsert' => [
+                    '$description' => 'This event triggers when a document is upserted.',
+                ],
+                'delete' => [
+                    '$description' => 'This event triggers when a row is deleted.'
+                ],
+            ],
+            'indexes' => [
+                '$model' => Response::MODEL_COLUMN_INDEX,
+                '$resource' => true,
+                '$description' => 'This event triggers on any indexes event.',
+                'create' => [
+                    '$description' => 'This event triggers when an index is created.',
+                ],
+                'update' => [
+                    '$description' => 'This event triggers when an index is updated.',
+                ],
+                'delete' => [
+                    '$description' => 'This event triggers when an index is deleted.'
+                ]
+            ],
+            'columns' => [
+                '$model' => Response::MODEL_COLUMN,
+                '$resource' => true,
+                '$description' => 'This event triggers on any columns event.',
+                'create' => [
+                    '$description' => 'This event triggers when a column is created.',
+                ],
+                'delete' => [
+                    '$description' => 'This event triggers when an column is deleted.'
+                ],
+                'update' => [
+                    '$description' => 'This event triggers when a column is created.',
+                ],
+            ],
+            'create' => [
+                '$description' => 'This event triggers when a table is created.'
+            ],
+            'update' => [
+                '$description' => 'This event triggers when a table is updated.',
+            ],
+            'delete' => [
+                '$description' => 'This event triggers when a table is deleted.',
+            ],
+        ],
         'collections' => [
             '$model' => Response::MODEL_COLLECTION,
             '$resource' => true,
@@ -106,11 +165,14 @@ return [
                 'create' => [
                     '$description' => 'This event triggers when a document is created.',
                 ],
-                'delete' => [
-                    '$description' => 'This event triggers when a document is deleted.'
-                ],
                 'update' => [
                     '$description' => 'This event triggers when a document is updated.'
+                ],
+                'upsert' => [
+                    '$description' => 'This event triggers when a document is upserted.',
+                ],
+                'delete' => [
+                    '$description' => 'This event triggers when a document is deleted.'
                 ],
             ],
             'indexes' => [
@@ -122,7 +184,10 @@ return [
                 ],
                 'delete' => [
                     '$description' => 'This event triggers when an index is deleted.'
-                ]
+                ],
+                'update' => [
+                    '$description' => 'This event triggers when a column is created.',
+                ],
             ],
             'attributes' => [
                 '$model' => Response::MODEL_ATTRIBUTE,
@@ -131,6 +196,9 @@ return [
                 'create' => [
                     '$description' => 'This event triggers when an attribute is created.',
                 ],
+                'update' => [
+                    '$description' => 'This event triggers when a column is created.',
+                ],
                 'delete' => [
                     '$description' => 'This event triggers when an attribute is deleted.'
                 ]
@@ -138,22 +206,22 @@ return [
             'create' => [
                 '$description' => 'This event triggers when a collection is created.'
             ],
+            'update' => [
+                '$description' => 'This event triggers when a collection is updated.',
+            ],
             'delete' => [
                 '$description' => 'This event triggers when a collection is deleted.',
             ],
-            'update' => [
-                '$description' => 'This event triggers when a collection is updated.',
-            ]
         ],
         'create' => [
             '$description' => 'This event triggers when a database is created.'
         ],
+        'update' => [
+            '$description' => 'This event triggers when a database is updated.',
+        ],
         'delete' => [
             '$description' => 'This event triggers when a database is deleted.',
         ],
-        'update' => [
-            '$description' => 'This event triggers when a database is updated.',
-        ]
     ],
     'buckets' => [
         '$model' => Response::MODEL_BUCKET,