+++ tests/extensions/Async/Exceptions/Critical.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Appwrite\Tests\Async\Exceptions;
+
+class Critical extends \Exception
+{
+}